It has come to my attention that some of the code from the RDA merge may have been leaked/stolen.
It is a violation of the license we have with Valve to use leaked or
stolen code, so five pieces of code have been removed, including the
client-side sprite implementation.

I have written a new client-side sprite implementation from scratch
based on C_ASW_Snow_Volume, so we don't have to lose any features as a
result of this.