AlterDeathMessage hook vscript function
called when a death message is about to be printed. Return a string which will be printed instead, %s1 in the string being the victim's display name (in case someone wants to do localization of their custom death message). No return value or non-string return value results in default death message