dont find entities which are marked for deletion
was leading to vscript crashes in a typical while find entity loop especially if ran every tick. entity would be marked for deletion and deemed invalid with IsValid function, but the find function would resurrect it in the same tick, then trying to use the resurrected handle shortly afterwards would crash the game. Also now doing Destroy in the typical while find loop is fine